refactor(SD-LEO-REFAC-EXTRACT-S17-ARCHETYPE-001): extract S17 refinement, delete legacy archetype generation
Decouple the LIVE GVOS Stage-17 refinement from the DEAD legacy archetype
generation so the generation-only code could finally be deleted (the deferral
from SD-LEO-REFAC-RETIRE-LEGACY-STAGE-001).

Phase 1 - extract:
- New lib/eva/stage-17/refinement.js holds generateRefinedVariants verbatim
  (only writeArtifact + dynamic getLLMClient; tokens stay a param).
- selection-flow.js repointed to ./refinement.js; new refinement.test.js.

Phase 2 - delete generation-only legacy (consumer-audited first):
- Removed archetype-generator.js + exclusive helpers (page-type-classifier,
  design-system-brief, content-brief-builder, scoring-engine).
- Removed dead entry points: server/routes/stage17.js /archetypes handler +
  isGvosComposerEnabled gate; worker _reconcileS17Archetypes + the legacy
  generateArchetypes block in _postStageHook_S17_DocGen (strategy gate kept);
  auto-resume.js + its server/index.js startup call.
- Removed the now-obsolete legacy tests.

KEPT (live consumers confirmed): design-mastering.js, strategy-stats.js,
strategy-recommender.js + the /strategy-recommendation route, token-manifest.js,
and the artifact-types S17 enum (schema contract).

Net -3014 LOC. Stage-17 unit suite delta = 0 (same 5 pre-existing
isDesignPassComplete mock failures, no new failures). grep-clean of
generateArchetypes outside the frozen audit snapshot.
