+++ lib/eva/stage-17/selection-flow.js
@@ -16,7 +16,7 @@
  * @module lib/eva/stage-17/selection-flow
  */
 
-import { generateRefinedVariants } from './archetype-generator.js';
+import { generateRefinedVariants } from './refinement.js';
 import { getTokenConstraints } from './token-manifest.js';
 import { writeArtifact } from '../artifact-persistence-service.js';
 import { masterDesign } from './design-mastering.js';
