+++ lib/eva/stage-execution-worker.js
@@ -435,7 +435,6 @@ export class StageExecutionWorker {
       );
       await this._releaseStaleLocks();
       await this._checkResolvedBlocks();
-      await this._reconcileS17Archetypes();
       await this._markStaleExecutions();
 
       const ventures = await this._pollForWork();
@@ -1884,115 +1883,6 @@ export class StageExecutionWorker {
   }
 
   /**
-   * S17 Archetype Reconciler — detect ventures at S17 with wireframe_screens
-   * but incomplete s17_archetypes and trigger generation.
-   *
-   * Runs every tick but rate-limited: skips ventures attempted in last 5 minutes.
-   * Uses stateless resume (generateArchetypes skips completed screens).
-   *
-   * SD-MAN-REFAC-S17-SIMPLIFY-PIPELINE-001: Systemic recovery for stuck S17 ventures.
-   */
-  async _reconcileS17Archetypes() {
-    try {
-      // Find active ventures at S17-S19 — PAT-RECONCILER-SCOPE-001:
-      // Reconciler must cover the full window where s17_archetypes are relevant.
-      // After chairman approval, _advanceStage moves the venture to S18+, but
-      // the fire-and-forget archetype generation may still be incomplete.
-      const { data: s17Ventures } = await this._supabase
-        .from('ventures')
-        .select('id, name')
-        .eq('status', 'active')
-        .gte('current_lifecycle_stage', 17)
-        .lte('current_lifecycle_stage', 19);
-
-      if (!s17Ventures || s17Ventures.length === 0) return;
-
-      for (const venture of s17Ventures) {
-        // Concurrency guard: skip if generation is already running for this venture
-        if (!this._s17ReconcileAttempts) this._s17ReconcileAttempts = new Map();
-        if (!this._s17ActiveGenerations) this._s17ActiveGenerations = new Set();
-        if (this._s17ActiveGenerations.has(venture.id)) continue;
-
-        // Rate limit: skip if attempted in last 30 minutes (generation takes 15-20 min)
-        const lastAttempt = this._s17ReconcileAttempts.get(venture.id);
-        if (lastAttempt && Date.now() - lastAttempt < 1_800_000) continue;
-
-        // Check if wireframe_screens exists
-        const { data: wireframes } = await this._supabase
-          .from('venture_artifacts')
-          .select('artifact_data')
-          .eq('venture_id', venture.id)
-          .eq('artifact_type', 'wireframe_screens')
-          .eq('is_current', true)
-          .limit(1)
-          .maybeSingle();
-
-        if (!wireframes?.artifact_data?.screens?.length) continue;
-
-        const expectedScreens = wireframes.artifact_data.screens.length;
-
-        // Count existing archetypes
-        const { count } = await this._supabase
-          .from('venture_artifacts')
-          .select('id', { count: 'exact', head: true })
-          .eq('venture_id', venture.id)
-          .eq('artifact_type', 's17_archetypes')
-          .eq('is_current', true);
-
-        if ((count ?? 0) >= expectedScreens) continue; // All done
-
-        // SD-S17-WORKER-STRATEGY-GATE-ORCH-001-A: Skip ventures without approved strategy decision
-        const { data: strategyDecision } = await this._supabase
-          .from('chairman_decisions')
-          .select('status, metadata')
-          .eq('venture_id', venture.id)
-          .eq('lifecycle_stage', 17)
-          .eq('metadata->>decision_type', 'strategy_selection')
-          .limit(1)
-          .maybeSingle();
-
-        if (!strategyDecision || strategyDecision.status !== 'approved') continue;
-        const reconStrategy = strategyDecision.metadata?.strategy || undefined;
-
-        // SD-LEO-FEAT-GVOS-ACTIVATION-REMEDIATION-001/FR-6: gate legacy archetype-generator
-        // behind s17_use_gvos_composer flag. When flag is ON for the venture, skip the
-        // legacy Opus-burning generation path — GvosS17Sections renders from venture_gvos_profile instead.
-        const { isEnabled: isFlagEnabledReconciler } = await import('../feature-flags/evaluator.js');
-        const useGvosComposerReconciler = await isFlagEnabledReconciler('s17_use_gvos_composer', { subjectId: venture.id }).catch(() => false);
-        if (useGvosComposerReconciler) {
-          this._logger.log(`[Worker] S17 reconciler: ${venture.name} — FLAG ON, skipping legacy archetype-generator`);
-          continue;
-        }
-
-        // Trigger generation
-        this._logger.log(`[Worker] S17 reconciler: ${venture.name} has ${count ?? 0}/${expectedScreens} archetypes — triggering generation (strategy: ${reconStrategy || 'all'})`);
-        this._s17ReconcileAttempts.set(venture.id, Date.now());
-        this._s17ActiveGenerations.add(venture.id);
-
-        // Fire-and-forget with concurrency tracking — don't block the tick loop
-        import('./stage-17/archetype-generator.js').then(({ generateArchetypes }) => {
-          generateArchetypes(venture.id, this._supabase, reconStrategy ? { strategy: reconStrategy } : undefined)
-            .then(result => {
-              this._logger.log(`[Worker] S17 reconciler: ${venture.name} complete — ${result.artifactIds.length} artifacts`);
-            })
-            .catch(err => {
-              this._logger.warn(`[Worker] S17 reconciler: ${venture.name} failed — ${err.message}`);
-            })
-            .finally(() => {
-              this._s17ActiveGenerations.delete(venture.id);
-            });
-        }).catch(importErr => {
-          this._logger.warn(`[Worker] S17 reconciler import error: ${importErr.message}`);
-          this._s17ActiveGenerations.delete(venture.id);
-        });
-      }
-    } catch (err) {
-      // Non-fatal — don't crash the tick loop
-      this._logger.warn(`[Worker] S17 reconciler error: ${err.message}`);
-    }
-  }
-
-  /**
    * Safety net: find ventures stuck in 'blocked' state whose blocking condition
    * has been resolved by an external process (e.g., rescan_stage_20 RPC).
    *
@@ -2855,32 +2745,12 @@ export class StageExecutionWorker {
     });
     this._logger.log('[Worker] S17 post-stage hook: vision + architecture docs generated');
 
-    // SD-S17-WORKER-STRATEGY-GATE-ORCH-001-A: Strategy gate — block until chairman selects strategy
-    // Before archetype generation, ensure strategy recommendation exists and is approved.
-    const selectedStrategy = await this._ensureS17StrategySelected(ventureId);
-
-    // SD-LEO-FEAT-GVOS-ACTIVATION-REMEDIATION-001/FR-6: gate legacy archetype-generator
-    // behind s17_use_gvos_composer flag. When flag is ON for the venture, skip the
-    // legacy Opus-burning generation path — GvosS17Sections renders from venture_gvos_profile.
-    const { isEnabled: isFlagEnabledS17 } = await import('../feature-flags/evaluator.js');
-    const useGvosComposerS17 = await isFlagEnabledS17('s17_use_gvos_composer', { subjectId: ventureId }).catch(() => false);
-    if (useGvosComposerS17) {
-      this._logger.log(`[Worker] S17 post-stage hook: FLAG ON for venture ${ventureId} — skipping legacy archetype-generator`);
-      return;
-    }
-
-    // SD-MAN-REFAC-S17-SIMPLIFY-FRONTEND-001: Auto-trigger archetype generation
-    // from the durable worker process (not fire-and-forget Express promise).
-    // Uses stateless resume — skips screens that already have completed artifacts.
-    try {
-      const { generateArchetypes } = await import('./stage-17/archetype-generator.js');
-      this._logger.log(`[Worker] S17 post-stage hook: starting archetype generation (strategy: ${selectedStrategy || 'all'})...`);
-      const result = await generateArchetypes(ventureId, this._supabase, selectedStrategy ? { strategy: selectedStrategy } : undefined);
-      this._logger.log(`[Worker] S17 archetypes complete: ${result.screenCount} screens, ${result.artifactIds.length} artifacts`);
-    } catch (archErr) {
-      // Non-fatal: chairman can re-trigger from frontend if worker generation fails
-      this._logger.warn(`[Worker] S17 archetype generation failed (non-fatal): ${archErr.message}`);
-    }
+    // SD-S17-WORKER-STRATEGY-GATE-ORCH-001-A: Strategy gate — block until chairman selects strategy.
+    // Required regardless of the design path: the GVOS composer renders S17 from
+    // venture_gvos_profile after strategy approval.
+    // SD-LEO-REFAC-EXTRACT-S17-ARCHETYPE-001: legacy archetype-generator invocation removed
+    // (GVOS composer is the live path; s17_use_gvos_composer is globally enabled).
+    await this._ensureS17StrategySelected(ventureId);
   }
 
   /**
